backend/app/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
# asyncpg==0.29.0   <-- You can also remove this, as it's not needed with psycopg2

# Validation & Security
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File Generation
reportlab==4.0.9
pandas==2.2.2
openpyxl==3.1.2

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional: Task Queue (keep if you plan to use Celery & Redis)
# celery==5.3.4
# redis==5.0.1